requirements.txt
Flask==3.0.0
pytest==7.4.0
python-dotenv
flask-cors
pytest-mock
mock
openai